pyproject.toml
[project]
name = "agno-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.21",
    "anthropic>=0.75.0",
    "fastapi>=0.128.0",
    "google>=3.0.0",
    "google-genai>=1.56.0",
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.30.0",
]
